+++ src/CleanRegex/Replace/Callback/ReplaceCallbackObject.php
@@ -13,7 +13,6 @@ use TRegx\CleanRegex\Match\Details\Detail;
 use TRegx\CleanRegex\Match\Details\Group\BaseDetailGroup;
 use TRegx\CleanRegex\Match\Details\MatchDetail;
 use TRegx\CleanRegex\Match\Details\ReplaceDetail;
-use TRegx\CleanRegex\Match\Details\ReplaceDetailImpl;
 use function call_user_func;
 use function is_string;
 
@@ -75,7 +74,7 @@ class ReplaceCallbackObject
     private function createDetailObject(): ReplaceDetail
     {
         $index = $this->counter++;
-        return new ReplaceDetailImpl(
+        return new ReplaceDetail(
             new MatchDetail(
                 $this->subject,
                 $index,